requirements.txt
grpcio
grpcio-tools
redis
requests
protobuf
